fix #2: leftover temp dir (zip); zip demo
Issue #2 on GitHub: if loading the carddeck as a zip the temporary
directory the zip was extracted into would be left, which would lead to
errors when the card deck was loaded again and it tried to create a
temporary directory to extract to and the directory already exists.
Resolved by both checking if the temporary directory already exists
first (and removing if it does) and also cleaning up the temporary
directory before quitting, basically

Since #2 was resolved I zipped up the demo carddeck (examples/xmas) to
an examples/xmas.carddeck.zip instead. I think this may encourage people
to easily download the demos separately.